Solution to Get Lowest Set Bit of a Number

+++ Get Lowest Set Bit of a Number@@67882a731b4392d8fc1577f0/code.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include<math.h>
 int main() {
     int num;
     scanf("%d",&num);
@@ -7,7 +8,7 @@ int main() {
         printf("-1\n");
         return 0;
     }
-    int position = log2(num & -num);
+    int position = log2((double)(num & -num));
     printf("%d\n",position);
     return 0;
 }
